samples/Server_Python/requirements.txt
tornado 
protobuf
kafka-python
